backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
feedparser==6.0.11
pydantic==1.10.18
python-dotenv==1.0.1
httpx==0.27.2
supabase==2.9.0
groq>=0.4.0
